[ADD] correcting-some-styles-and-a-confirmation

+++ js/controllers/applierVacanciesController.js
@@ -39,16 +39,30 @@ talenting.controller('applierVacanciesController', ['$scope', '$http', '$locatio
 
     $scope.disapply = (vacancyId, index) => {
 
-        $http({
-            method: "DELETE",
-            url: 'http://localhost:8080/talenting/appliersInVacancies',
-            params: {vacancyId: vacancyId}
-        }).then( () => {
 
-            $scope.appliedVacancies.splice(index, 1);
-
-        });
+        Swal.fire({
+            title: '¿Estas seguro?',
+            text: "Si retiras tu postulación todo tu proceso de contratación será cancelado",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#198754',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Aceptar',
+            cancelButtonText: 'Cancelar'
+        }).then((result) => {
+            if (result.isConfirmed) {
+                $http({
+                    method: "DELETE",
+                    url: 'http://localhost:8080/talenting/appliersInVacancies',
+                    params: {vacancyId: vacancyId}
+                }).then( () => {
+                    $scope.appliedVacancies.splice(index, 1);
+                    alertService.showAlert.info('¡Se ha retirado tu postulación!');
+                });
+            }
+        })
     };
+
 
 
 }]);
